combine unit test
